Initial version.
Includes preliminary Core-API with Interfaces and a generic Manager for any type of possible implementing subclass from external repositories.
Also includes an untested baseline implementation for XSLT- and RDF- transformation. Basic TSV/TTL Loaders and Writers to follow very soon.

Package structure and I/O for StreamComponents may still be subject to change.

This implementation is not yet ready for thorough testing but rather for internal discussion of the core classes and general structure.